Enable embedding waypoint data in maps.
This means Frogbot support can be added to a map without having to recompile the `qwprogs.dat`.
Embedded waypoints will override built-in (compiled) waypoints unless `sv_frog_only_builtin` is nonzero.

In theory one could manually add `FrB_` fields to entities in an editor like TrenchBroom, but next on the TODO list is to create a script that can inject data dumped from the waypoint tool into a `.map` file, which will be much more practical.