Fix rocket jumping.
The bots would almost never RJ because it relied almost entirely on chance.
RJ paths had an ‘infinite’ travel time that made bots generally ignore them, only in very specific cases was this overridden.
Now the path weights will vary depending on whether the bot is able (and willing) to RJ or not.
The mechanism in which this is implemented is rather… special.

Also try to make the jumps slightly more accurate.