Revise teleport paths
Disregard previous advice to never make incoming paths to teleport destinations, there are situations where it is necessary.
Fix this in all the maps where I messed it up.
Also tweak tree in `hohoho2` to give the bots early exits from the one-way path if they accidentally end up in it.