Deterministic marker handling in waypoint tool.
Stop relying on `find()` to gather markers for saving waypoints, it is unpredictable.
By ensuring marker indices are kept up-to-date, we can rely on linked lists instead.
Since we need a forward linked list for the `N()` section anyhow, also use it to dump paths etc in ascending order.
(Only teleport destinations break the nicely increasing count due to the way they are spawned.)
This makes it much easier to track changes.