Universal Shootable Triggers.
Replaces `dm6_door`, and extends it to any number of shootable triggers in any direction.

Some extra entities will now be available as markers, because they need to be targetable.
Updated all maps that used `dm6_door` or that can be extended with the new logic. Even the bridge in `cmt3` now works. 🤯